kern_resources_new/creative_lab/crew_ai/requirements.txt
# Requirements for CrewAI with GroqCloud

# Core dependencies
requests>=2.28.0
python-dotenv>=0.20.0

# CrewAI
crewai>=0.28.0

# Optional dependencies
# Uncomment if needed
# langchain>=0.0.267
# openai>=1.3.0
# anthropic>=0.5.0
